test(autopipeline): consolidate full-duplex tests into one file
Merge the seven mergeable full-duplex test files into
autopipeline_fullduplex_test.go and delete the sources: the four
autopipeline_fd_reviewfixes*_test.go files plus autopipeline_fdpool/fdring/fdsparse.
Pure move; 70 Test funcs preserved, no duplicate declarations. The two //go:build
fuzz files stay separate (they cannot fold into an untagged file). Also swaps the
obsolete hook-snapshot test for TestFDLiveHookChain, matching the snapshot removal in
the next commit.
